+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # torrt changelog
 
+### Unreleased
+* ** qBittorrent: preserve torrent category on update.
+
 ### v1.1.0 [2026-03-30]
 * !! QA. Dropped QA for Py 3.7, 3.8, 3.9, 3.10.
 * ++ Add support to pass custom RPC parameters.